WIP: Decimate caching layer
It turns out that trying to implement a complicated caching model
is more trouble than it's worth. For instance, dealing with the
versioning of nested objects is difficult with our current
ETag scheme.

This commit (mostly) removes the old caching layer; we will then
add back GeoPackage-based caching for views.